- Updates orc8r libs to model the fact that all services will offer an interface that displays if their service is active. - Metricsd no longer relies directly on service-registry with the intent of leveraging those new changes in orc8r libs.
