+++ orchestrator-bundle/orc8r-libs/lib/charms/magma_orc8r_libs/v0/orc8r_base_db.py
@@ -20,7 +20,6 @@ ops-lib-pgsql
 EOF
 ```
 Then, to initialise the library:
-For ClusterIP services:
 ```python
 from charms.magma_orc8r_libs.v0.orc8r_base_db import Orc8rBase
 from charms.observability_libs.v0.kubernetes_service_patch import KubernetesServicePatch
@@ -31,7 +30,7 @@ from ops.main import main
 class MagmaOrc8rDirectorydCharm(CharmBase):
     def __init__(self, *args):
         super().__init__(*args)
-        self._service_patcher = KubernetesServicePatch(self, [("grpc", 9180, 9106)])
+        self._service_patcher = KubernetesServicePatch(self, [("grpc", 9180)])
         startup_command = (
             "/usr/bin/envdir "
             "/var/opt/magma/envdir "
@@ -42,6 +41,16 @@ class MagmaOrc8rDirectorydCharm(CharmBase):
         self._orc8r_base = Orc8rBase(self, startup_command=startup_command)
 
 ```
+
+Charms that leverage this library also need to specify a `provides` relation in their
+`metadata.yaml` file. For example:
+
+```yaml
+provides:
+  magma-orc8r-directoryd:
+    interface: magma-orc8r-directoryd
+```
+
 """
 
 import logging
@@ -49,7 +58,7 @@ import logging
 import ops.lib
 from ops.charm import CharmBase
 from ops.framework import Object
-from ops.model import ActiveStatus, BlockedStatus, MaintenanceStatus, WaitingStatus
+from ops.model import ActiveStatus, BlockedStatus, MaintenanceStatus, WaitingStatus, ModelError
 from ops.pebble import Layer
 from pgconnstr import ConnectionString  # type: ignore[import]
 
@@ -80,11 +89,13 @@ class Orc8rBase(Object):
         super().__init__(charm, "orc8r-base")
         self.charm = charm
         self.startup_command = startup_command
-        self._container_name = self._service_name = self.charm.meta.name
-        self.container = self.charm.unit.get_container(self._container_name)
+        self.container_name = self.service_name = self.charm.meta.name
+        self.container = self.charm.unit.get_container(self.container_name)
+        name_with_underscores = self.service_name.replace('-', '_')
         pebble_ready_event = getattr(
-            self.charm.on, f"{self._service_name.replace('-', '_')}_pebble_ready"
+            self.charm.on, f"{name_with_underscores}_pebble_ready"
         )
+        relation_joined_event = getattr(self.charm.on, f"{name_with_underscores}_relation_joined")
         self.framework.observe(pebble_ready_event, self._on_magma_orc8r_pebble_ready)
 
         if additional_environment_variables:
@@ -92,10 +103,11 @@ class Orc8rBase(Object):
         else:
             self.additional_environment_variables = {}
 
-        self._db = pgsql.PostgreSQLClient(self.charm, "db")
+        self.db = pgsql.PostgreSQLClient(self.charm, "db")
         self.framework.observe(
-            self._db.on.database_relation_joined, self._on_database_relation_joined
+            self.db.on.database_relation_joined, self._on_database_relation_joined
         )
+        self.framework.observe(relation_joined_event, self._on_relation_joined)
 
     @property
     def _db_relation_created(self) -> bool:
@@ -109,7 +121,7 @@ class Orc8rBase(Object):
             self.charm.unit.status = BlockedStatus("Waiting for database relation to be created")
             event.defer()
             return
-        if not self._db_relation_established:
+        if not self._db_relation_created:
             self.charm.unit.status = WaitingStatus(
                 "Waiting for database relation to be established..."
             )
@@ -126,10 +138,11 @@ class Orc8rBase(Object):
             pebble_layer = self._pebble_layer()
             plan = self.container.get_plan()
             if plan.services != pebble_layer.services:
-                self.container.add_layer(self._container_name, pebble_layer, combine=True)
-                self.container.restart(self._service_name)
-                logger.info(f"Restarted container {self._service_name}")
-                self.charm.unit.status = ActiveStatus()
+                self.container.add_layer(self.container_name, pebble_layer, combine=True)
+            self.container.restart(self.service_name)
+            logger.info(f"Restarted container {self.service_name}")
+            self._update_relations()
+            self.charm.unit.status = ActiveStatus()
         else:
             self.charm.unit.status = WaitingStatus("Waiting for container to be ready...")
             event.defer()
@@ -138,12 +151,12 @@ class Orc8rBase(Object):
         """Returns pebble layer for the charm."""
         return Layer(
             {
-                "summary": f"{self._service_name} layer",
-                "description": f"pebble config layer for {self._service_name}",
+                "summary": f"{self.service_name} layer",
+                "description": f"pebble config layer for {self.service_name}",
                 "services": {
-                    self._service_name: {
+                    self.service_name: {
                         "override": "replace",
-                        "summary": self._service_name,
+                        "summary": self.service_name,
                         "startup": "enabled",
                         "command": self.startup_command,
                         "environment": self._environment_variables,
@@ -187,9 +200,9 @@ class Orc8rBase(Object):
     def _environment_variables(self):
         environment_variables = {}
         default_environment_variables = {
-            "SERVICE_HOSTNAME": self._container_name,
+            "SERVICE_HOSTNAME": self.container_name,
             "SERVICE_REGISTRY_MODE": "k8s",
-            "SERVICE_REGISTRY_NAMESPACE": self._namespace,
+            "SERVICE_REGISTRY_NAMESPACE": self.namespace,
         }
         environment_variables.update(self.additional_environment_variables)
         environment_variables.update(default_environment_variables)
@@ -201,11 +214,46 @@ class Orc8rBase(Object):
             f"sslmode=disable",
             "SQL_DRIVER": "postgres",
             "SQL_DIALECT": "psql",
-            "SERVICE_HOSTNAME": self._container_name,
+            "SERVICE_HOSTNAME": self.container_name,
         }
         environment_variables.update(sql_environment_variables)
         return environment_variables
 
     @property
-    def _namespace(self) -> str:
+    def namespace(self) -> str:
         return self.charm.model.name
+
+    def _update_relations(self):
+        if not self.charm.unit.is_leader():
+            return
+        relations = self.charm.model.relations[self.charm.meta.name]
+        for relation in relations:
+            self._update_relation_active_status(
+                relation=relation,
+                is_active=self._service_is_running
+            )
+
+    def _on_relation_joined(self, event):
+        if not self.charm.unit.is_leader():
+            return
+        self._update_relation_active_status(
+            relation=event.relation,
+            is_active=self._service_is_running
+        )
+
+    @property
+    def _service_is_running(self) -> bool:
+        if self.container.can_connect():
+            try:
+                self.container.get_service(self.service_name)
+                return True
+            except ModelError:
+                pass
+        return False
+
+    def _update_relation_active_status(self, relation, is_active: bool):
+        relation.data[self.charm.unit].update(
+            {
+                "active": str(is_active),
+            }
+        )
