+++ orchestrator-bundle/orc8r-metricsd-operator/lib/charms/magma_orc8r_libs/v0/orc8r_base.py
@@ -20,6 +20,7 @@ charmcraft fetch-lib charms.magma_orc8r_libs.v0.orc8r_base
 ```
 
 Then, to initialise the library:
+For ClusterIP services:
 
 ```python
 
@@ -32,7 +33,7 @@ from ops.main import main
 class MagmaOrc8rHACharm(CharmBase):
     def __init__(self, *args):
         super().__init__(*args)
-        self._service_patcher = KubernetesServicePatch(self, [("grpc", 9180)])
+        self._service_patcher = KubernetesServicePatch(self, [("grpc", 9180, 9119)])
         startup_command = (
             "/usr/bin/envdir "
             "/var/opt/magma/envdir "
@@ -43,15 +44,6 @@ class MagmaOrc8rHACharm(CharmBase):
         self._orc8r_base = Orc8rBase(self, startup_command=startup_command)
 ```
 
-Charms that leverage this library also need to specify a `provides` relation in their
-`metadata.yaml` file. For example:
-
-```yaml
-provides:
-  magma-orc8r-ha:
-    interface: magma-orc8r-ha
-```
-
 """
 
 
@@ -59,7 +51,7 @@ import logging
 
 from ops.charm import CharmBase
 from ops.framework import Object
-from ops.model import ActiveStatus, MaintenanceStatus, WaitingStatus, ModelError
+from ops.model import ActiveStatus, MaintenanceStatus, WaitingStatus
 from ops.pebble import Layer
 
 # The unique Charmhub library identifier, never change it
@@ -70,7 +62,7 @@ LIBAPI = 0
 
 # Increment this PATCH version before using `charmcraft publish-lib` or reset
 # to 0 if you are raising the major API version
-LIBPATCH = 9
+LIBPATCH = 8
 
 
 logger = logging.getLogger(__name__)
@@ -86,13 +78,12 @@ class Orc8rBase(Object):
         super().__init__(charm, "orc8r-base")
         self.charm = charm
         self.startup_command = startup_command
-        self.container_name = self.service_name = self.charm.meta.name
-        name_with_underscores = self.service_name.replace('-', '_')
-        pebble_ready_event = getattr(self.charm.on, f"{name_with_underscores}_pebble_ready")
-        relation_joined_event = getattr(self.charm.on, f"{name_with_underscores}_relation_joined")
-        self._container = self.charm.unit.get_container(self.container_name)
+        self._container_name = self._service_name = self.charm.meta.name
+        pebble_ready_event = getattr(
+            self.charm.on, f"{self._service_name.replace('-', '_')}_pebble_ready"
+        )
+        self._container = self.charm.unit.get_container(self._container_name)
         self.framework.observe(pebble_ready_event, self._on_magma_orc8r_pebble_ready)
-        self.framework.observe(relation_joined_event, self._on_relation_joined)
 
         if additional_environment_variables:
             self.additional_environment_variables = additional_environment_variables
@@ -111,11 +102,10 @@ class Orc8rBase(Object):
             pebble_layer = self._pebble_layer()
             plan = self._container.get_plan()
             if plan.services != pebble_layer.services:
-                self._container.add_layer(self.container_name, pebble_layer, combine=True)
-            self._container.restart(self.service_name)
-            logger.info(f"Restarted container {self.service_name}")
-            self._update_relations()
-            self.charm.unit.status = ActiveStatus()
+                self._container.add_layer(self._container_name, pebble_layer, combine=True)
+                self._container.restart(self._service_name)
+                logger.info(f"Restarted container {self._service_name}")
+                self.charm.unit.status = ActiveStatus()
         else:
             self.charm.unit.status = WaitingStatus("Waiting for container to be ready...")
             event.defer()
@@ -124,12 +114,12 @@ class Orc8rBase(Object):
         """Returns pebble layer for the charm."""
         return Layer(
             {
-                "summary": f"{self.service_name} layer",
-                "description": f"pebble config layer for {self.service_name}",
+                "summary": f"{self._service_name} layer",
+                "description": f"pebble config layer for {self._service_name}",
                 "services": {
-                    self.service_name: {
+                    self._service_name: {
                         "override": "replace",
-                        "summary": self.service_name,
+                        "summary": self._service_name,
                         "startup": "enabled",
                         "command": self.startup_command,
                         "environment": self._environment_variables,
@@ -142,49 +132,14 @@ class Orc8rBase(Object):
     def _environment_variables(self):
         environment_variables = {}
         default_environment_variables = {
-            "SERVICE_HOSTNAME": self.container_name,
+            "SERVICE_HOSTNAME": self._container_name,
             "SERVICE_REGISTRY_MODE": "k8s",
-            "SERVICE_REGISTRY_NAMESPACE": self.namespace,
+            "SERVICE_REGISTRY_NAMESPACE": self._namespace,
         }
         environment_variables.update(self.additional_environment_variables)
         environment_variables.update(default_environment_variables)
         return environment_variables
 
     @property
-    def namespace(self) -> str:
+    def _namespace(self) -> str:
         return self.charm.model.name
-
-    def _update_relations(self):
-        if not self.charm.unit.is_leader():
-            return
-        relations = self.charm.model.relations[self.charm.meta.name]
-        for relation in relations:
-            self._update_relation_active_status(
-                relation=relation,
-                is_active=self._service_is_running
-            )
-
-    def _on_relation_joined(self, event):
-        if not self.charm.unit.is_leader():
-            return
-        self._update_relation_active_status(
-            relation=event.relation,
-            is_active=self._service_is_running
-        )
-
-    @property
-    def _service_is_running(self) -> bool:
-        if self._container.can_connect():
-            try:
-                self._container.get_service(self.service_name)
-                return True
-            except ModelError:
-                pass
-        return False
-
-    def _update_relation_active_status(self, relation, is_active: bool):
-        relation.data[self.charm.unit].update(
-            {
-                "active": str(is_active),
-            }
-        )
