+++ orchestrator-bundle/orc8r-metricsd-operator/metadata.yaml
@@ -25,7 +25,3 @@ storage:
     type: filesystem
     description: Configs storage
     minimum-size: 1M
-
-requires:
-  magma-service-registry:
-    interface: magma-service-registry
