+++ orchestrator-bundle/orc8r-service-registry-operator/metadata.yaml
@@ -18,7 +18,3 @@ resources:
     type: oci-image
     description: OCI image for magma-orc8r-service-registry (docker.artifactory.magmacore.org/controller:1.6.0)
     upstream-source: docker.artifactory.magmacore.org/controller:1.6.0
-
-provides:
-  magma-service-registry:
-    interface: magma-service-registry
